Upload

+++ controllers/admin.js
@@ -1,7 +1,8 @@
-const mongoose = require("mongoose");
 const { ValidationError } = require("sequelize");
+const fileHelper = require("../util/file");
 const Product = require("../models/product");
 const { validationResult } = require("express-validator");
+const mongoose = require("mongoose");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -16,26 +17,37 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.file;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  console.log(imageUrl);
   const errors = validationResult(req);
 
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/edit-product",
+      editing: false,
+      hasError: true,
+      product: { title, price, description },
+      errorMessage: "Attach File not an Image",
+      validationErrors: [],
+    });
+  }
+  const imageUrl = image.path;
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/edit-product",
       editing: false,
       hasError: true,
-      product: { title, imageUrl, price, description },
+      product: { title, price, description },
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array(),
     });
   }
 
   const product = new Product({
-    //_id: new mongoose.Types.ObjectId("668b72f12c5036332882574c"),
+    //_id: mongoose.Types.ObjectId("6687bd6c575ba15124ca476e"),
     title: title,
     price: price,
     description: description,
@@ -49,7 +61,8 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      // return res.status(500).render("admin/edit-product", {
+      //console.log(err);
+      // res.status(500).render("admin/edit-product", {
       //   pageTitle: "Add Product",
       //   path: "/admin/add-product",
       //   editing: false,
@@ -58,10 +71,7 @@ exports.postAddProduct = (req, res, next) => {
       //   errorMessage: "Database operation failed, please try again.",
       //   validationErrors: [],
       // });
-      //res.redirect("/500");
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
+      res.redirect("/500");
     });
 };
 
@@ -89,7 +99,7 @@ exports.getEditProduct = (req, res, next) => {
     .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
-      return next(error);
+      return next();
     });
 };
 
@@ -97,7 +107,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -128,7 +138,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl); //delete image on folder
+        product.imageUrl = image.path;
+      }
       return product.save();
     })
     .then((result) => {
@@ -177,13 +190,23 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        throw new Error("Product not found");
+      }
+      fileHelper.deleteFile(product.imageUrl); //delete image on folder
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
+    .catch((err) => next(err))
+
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
     })
     .catch((err) => {
-      console.log(err);
-      res.status(500).redirect("/admin/products");
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next();
     });
 };
